Rename v2.py to cal_v2.py
This version introduces the function solution method, which can output the function of the relationship between each solid phase and the corresponding ion concentration, and can plot the function (completed, not uploaded). However, this version is easy to find a solution due to the initial value problem, so it is necessary to draw a diagram first or use version 1.0 to calculate the approximate solution.